refactor: gruping rootlayout on one folder
