+++ client/src/components/RootComponents/NavBar.tsx
