+++ client/src/components/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as LeftSideBar } from './RootComponents/LeftSideBar';
+export { default as RightSideBar } from './RootComponents/RightSideBar';
